INGK-304 Follow typing style guide

+++ include/public/ingenialink/err.h
@@ -58,7 +58,7 @@ IL_BEGIN_DECL
  */
 IL_EXPORT const char *ilerr_last(void);
 IL_EXPORT int *ilerr_ipb_last(void);
-IL_EXPORT void set_log_level(int level);
+IL_EXPORT void set_log_level(int8_t level);
 
 /** @} */
