+++ ingenialink/err.c
@@ -151,7 +151,7 @@ int *ilerr_ipb_last()
 	return err_ipb_last;
 }
 
-void set_log_level(int level)
+void set_log_level(int8_t level)
 {
 	log_set_level(level);
 }
